+++ CHANGELOG.md
@@ -1,8 +1,5 @@
 # WaterDrop changelog
 
-## 0.3.3
-- Added `topic_prefix` config flag
-
 ## 0.3.2
 - Removed support for Ruby 2.1.*
 - ~~Ruby 2.3.3 as default~~
@@ -11,6 +8,7 @@
 - Dry configurable config (#20)
 - added .rspec for default spec helper require
 - Added SSL capabilities
+- Added `topic_prefix` config flag
 
 ## 0.3.1
 - Dev tools update